% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as_gt.wink_banner}
\alias{as_gt.wink_banner}
\title{Coerce wink_banner to gt_tbl
Defines how a 'wink_banner' object is converted into a fully
styled gt table for display in R Markdown, Shiny, or documents}
\usage{
\method{as_gt}{wink_banner}(data, ...)
}
\arguments{
\item{data}{An object of class 'wink_banner'.}

\item{...}{Not used.}
}
\value{
A gt_tbl object, fully styled with spanners, formatting, and notes.
}
\description{
Coerce wink_banner to gt_tbl
Defines how a 'wink_banner' object is converted into a fully
styled gt table for display in R Markdown, Shiny, or documents
}
