% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{Create banner crosstabs and apply signficance testing across groups}
\usage{
banner(
  data,
  var,
  bys,
  weight = NULL,
  var_nets = NULL,
  digits = 0,
  min_group_n = 100,
  exclude_var = NULL,
  exclude_bys = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{data}{(data.frame) A data frame of survey respondents}

\item{var}{(chr) the target variable (i.e. dependent variable)}

\item{weight}{(chr) A weight variable. If NULL, produces unweighted estimates}

\item{var_nets}{(list) Each element is either a character vector of factor levels of \code{var}, or a vector of integers, which are the converted to factor levels of \code{var}}

\item{digits}{(int) How many decimal places to round all numeric values to. NULL results in the default behavior of \code{pewmethods::get_totals} (i.e. many decimal places).}

\item{min_group_n}{(int) n-size threshold to be flagged within \code{by} groups.}

\item{exclude_var}{(chr) a regex - matching levels are excluded from \code{var} rows. If NULL, removes nothing}

\item{na.rm}{(lgl) Whether to remove missing values from \code{var} before calculating values}

\item{by}{(chr) The variable to be cross-tabulated with \code{var} (i.e. independent variable)}

\item{st_col_start}{(int) the letter of the alphabet used as the first label for significance testing comparisons}

\item{exclude_by}{(chr) a regex - matching levels are excluded from \code{by} columns. If NULL, removes nothing}
}
\value{
A data frame, where rows represent levels of \code{var} and columns represent levels of \code{bys}, as well as
the total distribution across the entire sample in the front. Each column contains the distribution of respones
to \code{var}, conditional on each \code{by} level.
}
\description{
Generates multiple crosstabs of variables \code{bys} against a target variable \code{var},
applying significance testing across columns of each \code{by} and combining the results into a single table
}
\details{
\code{exclude_bys} must have length 1 or \code{length(bys)}. If length == \code{length(bys)}, then each
pattern in \code{exclude_bys} is applied to each by of \code{bys} in turn. If length == 1, then
the same pattern is applied to all \code{bys}. All other settings throw an error.

In the case that \code{var} and \code{by} are the same, the result might include extra padding rows corresponding to any
levels of \code{var/by} that were empty, or if \code{na.rm = FALSE} is set and \code{var/by}'s missing values resulted in an
extra row for "(Missing)". In both cases, padding rows are added to ensure the Total columns and subsequent
crosstabs align properly.
}
\examples{
\dontrun{

# `food` is a dataset provided in the package.
banners(
  data = food,
  var = "vegan",
  bys = c("rating_meat", "rating_sushi", "rating_veg")
)
}
}
\seealso{
\code{\link{crosstab}} for the underlying crosstab function
}
