% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab2.R
\name{crosstab2}
\alias{crosstab2}
\title{Rewriting the crosstab function}
\usage{
crosstab2(
  data,
  var,
  by,
  weight = NULL,
  digits = 0,
  min_group_n = 100,
  st_col_start = 2
)
}
\description{
Rewriting the crosstab function
}
