% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab2.R
\name{crosstab_builder}
\alias{crosstab_builder}
\title{Building a crosstab from its inputs}
\usage{
crosstab_builder(
  baby_crosstab,
  sig_test_table,
  n_unweighted_row,
  deff_row,
  mose_row
)
}
\description{
Building a crosstab from its inputs
}
