% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{double_or_not}
\alias{double_or_not}
\title{safely coerce something to a double, or not}
\usage{
double_or_not(x)
}
\description{
Might use this to write values to excel sheets as numbers instead of text, when necessary
}
