% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_banners.R
\name{fix_guts}
\alias{fix_guts}
\title{Get the guts of a banner}
\usage{
fix_guts(data)
}
\description{
An internal function (used in \code{write_banners()}) that facilitates printing banners to excel such that
numbers are saved as numbers and numbers are saved as text.
}
