% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_totals.R
\name{get_totals}
\alias{get_totals}
\title{Get weighted percentages or totals}
\usage{
get_totals(
  data,
  var,
  by = NULL,
  wt = NULL,
  digits = NULL,
  complete = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{The \code{data.frame} containing the variables to be tabulated.}

\item{var}{\code{character}, indicating the name of the variable to be tabulated}

\item{by}{For creating cross-tabulations, an optional \code{character} variable
containing the name of the variable to be crossed with \code{var}. Can pass
multiple variables in as a vector.
Defaults to \code{FALSE}.}

\item{wt}{A \code{character} vector containing the name(s) of the weight variable(s)
to be used in tabulating results. If nothing is passed, results will be unweighted.}

\item{digits}{The number of decimal points displayed. Defaults to value specified
in \code{options("digits")}.}

\item{complete}{TRUE/FALSE: Should empty levels of \code{var} and \code{by} be
included in the results? Defaults to \code{TRUE}.}

\item{na.rm}{Should \code{NA} values of \code{var} and \code{by} be included as their own
levels? If \code{TRUE}, columns and rows for \code{"(Missing)"} values appear. If FALSE,
any \code{NA} values in \code{var} are excluded from the denominator of percentages, and
any \code{"(Missing)"} column for \code{by} is dropped}
}
\value{
A \code{data.frame} with a column for the levels of \code{var}, along with additional
columns of percentages, as specified by arguments to this function.
}
\description{
Takes a categorical variable and returns the weighted or unweighted
percentage or total of each category. Can include a grouping variable. Can be
used to compare weights versus one another.
}
\details{
If no arguments are supplied to \code{by}, then the output has one
column named "total". If arguments are supplied to \code{by}, then the
column names will be the categories of the grouping variable.
}
