% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.wink_banner}
\alias{print.wink_banner}
\title{Print method for 'wink_banner' class}
\usage{
\method{print}{wink_banner}(data, ...)
}
\arguments{
\item{data}{An object of class 'wink_banner'.}

\item{...}{Additional arguments (can ignore for now)}
}
\value{
The input object 'data' (invisibly)
}
\description{
Implements the custom print behavior for 'wink_banner' tables
}
