% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab2.R
\name{sig_test_table}
\alias{sig_test_table}
\title{Do significance testing on a crosstab}
\usage{
sig_test_table(baby_crosstab, n_unweighteds, deffs, col_start = 2)
}
\description{
Isolating the functionality related to significance testing, for easier maintenance
}
