% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{sig_test_table2}
\alias{sig_test_table2}
\title{Do significance testing on a crosstab}
\usage{
sig_test_table2(baby_crosstab, n_unweighteds, deffs, col_start = 3)
}
\description{
Isolating the functionality related to significance testing, for easier maintenance
}
