% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wink-package.R
\docType{package}
\name{wink-package}
\alias{wink}
\alias{wink-package}
\title{wink: Crosstab and Signifiance Testing for Weighted Survey Estimates (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Crosstabs and banners of weighted (and unweighted) survey estimates, exported to Excel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjmoncus.github.io/wink/}
}

}
\author{
\strong{Maintainer}: J.J. Moncus \email{jjmoncus706@gmail.com}

}
\keyword{internal}
