% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_banners.R
\name{write_banners}
\alias{write_banners}
\title{Export banners to a n excel workbook}
\usage{
write_banners(banners, file, overwrite = TRUE, format_numbers = TRUE)
}
\arguments{
\item{banners}{A banner or list of banners, generated with  \code{\link{banner}}}

\item{file}{(chr) The file path to save the output to. Should include the ".xlsx" extension}

\item{overwrite}{(lgl) Whether to overwrite an existing file. Default is TRUE. If FALSE, will
throw an error a file with the same name already exists.}

\item{format_numbers}{(lgl) Whether to format numeric values (percentages, n-sizes, etc). per default
styling. If TRUE, percentages and n-sizes are formatted to integer values, and design effects and margins of error
are formatted to two decimal places.}
}
\value{
Invisibly returns NULL. The function is called for its side effect of creating and writing
an Excel file.

The Excel file has several properties:
\itemize{
\item \strong{Table of Contents}: A dedicated first sheet with hyperlinks to all banners. Each banner also includes a button back to the table of contents.
\item \strong{worksheets}: Each banner is written to a separate sheet, named after the \code{var} variable used to generate it.
\item \strong{formatting}: Borders are added to separate sections of the table. Numbers are saved and formatted per rules, and text is saved as text.
}
}
\description{
Takes banners (generated with \code{banner()}) and prints them to a
formatted excel workbook. Each banner is printed to a separate worksheet with appropriate
styling, including borders and column dividers. A Table of Contents sheet with
hyperlinks is added to the front.
}
\examples{
\dontrun{

# Make some banners
# `food` is a dataset provided in the package
results <- banners(
  data = food,
  target = "vegan",
  banners = c("rating_meat", "rating_sushi")
)

# Export to Excel with default settings
write_banners(
  banners = results,
  file = "results.xlsx",
  overwrite = TRUE,
  format_numbers = TRUE
)
}

}
\seealso{
\code{\link{banners}} for creating banner analysis results,
\code{\link[openxlsx]{createWorkbook}} for Excel workbook creation,
\code{\link[openxlsx]{addWorksheet}} for adding worksheets,
\code{\link[openxlsx]{saveWorkbook}} for saving Excel files
}
