% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Rewriting the crosstab function}
\usage{
crosstab(
  data,
  var,
  by,
  weight = NULL,
  var_nets = NULL,
  digits = 0,
  min_group_n = 100,
  st_col_start = 3,
  exclude_var = NULL,
  exclude_by = NULL
)
}
\description{
Rewriting the crosstab function
}
\details{
When specifying \code{var_nets} using numbers instead of characters, remember to specify integers with reference to the \emph{original factor levels of the variable},
as opposed to the number-row of the table you think each level corresponds to. Interactions with \code{exclude_var} and \code{exclude_by}, might result
in number-rows no longer matching the original levels
}
