% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Rewriting the crosstab function}
\usage{
crosstab(
  data,
  var,
  by,
  weight = NULL,
  var_nets = NULL,
  digits = 0,
  min_group_n = 100,
  st_col_start = 3
)
}
\description{
Rewriting the crosstab function
}
