% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hardcode_sig_test}
\alias{hardcode_sig_test}
\title{hardcoded test for significant differences between two percentages}
\usage{
hardcode_sig_test(p1s, n1, deff1, p2s, n2, deff2)
}
\description{
hardcoded test for significant differences between two percentages
}
