% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wink-package.R
\docType{package}
\name{wink-package}
\alias{wink}
\alias{wink-package}
\title{wink: Crosstabs, Banners and Signifiance Testing for Survey Estimates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Banners of survey estimates, exportable to Excel for easy viewing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjmoncus.github.io/wink/}
}

}
\author{
\strong{Maintainer}: J.J. Moncus \email{jjmoncus706@gmail.com}

}
\keyword{internal}
