% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wink-package.R
\docType{package}
\name{wink-package}
\alias{wink}
\alias{wink-package}
\title{wink: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jjmoncus.github.io/wink/}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
